Refine property grouping (#382)
The updated grouping keeps most properties in the same order as before, just a bit more organised. However, there are a few changes to the order, particularly 
- `text-overflow-*` (now grouped with overflow)
- `text-decoration-*` (now grouped after basic text properties)
- `list-style-*` (now grouped after text properties)
- `outline-*` (now grouped before backgrounds and borders)
- `transition-*` (now grouped after transform)

The updated `groups.js` includes some placeholder comments for upcoming groups which gives these changes some additional context.

Not adding any new properties in this version in order to make the changes to the order easier to track/notice.

See #381